Unfix typography before creating local definition text.
