+++ bikeshed/__init__.py
@@ -180,21 +180,6 @@ def replaceTextMacros(text):
     return text
 
 
-def fixTypography(text):
-    # Replace straight aposes with curly quotes for possessives and contractions.
-    text = re.sub(r"([\w])'([\w])", r"\1’\2", text)
-    text = re.sub(r"(</[\w]+>)'([\w])", r"\1’\2", text)
-    # Fix line-ending em dashes, or --, by moving the previous line up, so no space.
-    text = re.sub(r"([^<][^!])(—|--)\r?\n\s+(\S)", r"\1—<wbr>\3", text)
-    return text
-
-def unfixTypography(text):
-    # Replace curly quotes with straight quotes, and emdashes with double dashes.
-    text = re.sub(r"’", r"'", text)
-    # Fix line-ending em dashes, or --, by moving the previous line up, so no space.
-    text = re.sub(r"—<wbr>", r"--", text)
-    return text
-
 def transformMarkdownParagraphs(doc):
     doc.lines = markdown.parse(doc.lines)
 
@@ -1032,7 +1017,7 @@ def processAutolinks(doc):
             el.set('href', '#'+simplifyText(linkText))
             continue
 
-        url = doc.refs.getRef(linkType, unfixTypography(linkText),
+        url = doc.refs.getRef(linkType, linkText,
                               spec=el.get('data-link-spec'),
                               status=el.get('data-link-status'),
                               linkFor=el.get('data-link-for'),
@@ -1144,7 +1129,7 @@ def processIDL(doc):
         for el in findAll("idl", pre):
             idlType = el.get('data-idl-type')
             idlText = el.get('title')
-            url = doc.refs.getRef(idlType, unfixTypography(idlText.lower()),
+            url = doc.refs.getRef(idlType, idlText.lower(),
                                   linkFor=el.get('data-idl-for'),
                                   el=el,
                                   error=False)
